settings animation duration

+++ src/main/java/bvb/core/BVBSettings.java
@@ -13,7 +13,7 @@ public class BVBSettings
 	public static boolean bShowVolumeBoxes = Prefs.get("BVB.bShowVolumeBoxes", true);
 
 	/** animation speed during zoomin in/out , i.e. duration of displayed transform in ms **/
-	public static long nTransformAnimationDuration =  (int)Prefs.get("BVB.nTransformAnimationDuration",400);
+	public static int nTransformAnimationDuration =  (int)Prefs.get("BVB.nTransformAnimationDuration",400);
 		
 	/** IO default or last folder **/	
 	public static String lastDir = Prefs.get( "BVB.lastDir", "" );
